added tests Cleanup

+++ lib/services/node_service.dart
@@ -1,4 +1,6 @@
+import 'package:kiss_repository/kiss_repository.dart';
 import 'package:uuid/uuid.dart';
+
 import '../models/node.dart';
 import '../repositories/node_repository.dart';
 
@@ -11,7 +13,7 @@ class NodeService {
   Future<Node> createNode(NodeCreate nodeCreate) async {
     String id = _uuid.v4();
     String rootId = id;
-    
+
     if (nodeCreate.previous != null) {
       try {
         final parentNode = await _repository.get(nodeCreate.previous!);
@@ -20,7 +22,7 @@ class NodeService {
         throw Exception('Parent node not found: ${nodeCreate.previous}');
       }
     }
-    
+
     final node = Node(
       id: id,
       root: rootId,
@@ -28,8 +30,8 @@ class NodeService {
       spatialHash: nodeCreate.spatialHash,
       content: nodeCreate.content,
     );
-    
-    return await _repository.add(node);
+
+    return await _repository.addNode(node);
   }
 
   Future<Node> getNode(String id) async {
@@ -46,6 +48,15 @@ class NodeService {
   }
 
   Future<void> deleteNode(String id) async {
+    // Business logic: prevent deletion of nodes with children
+    final children = await _repository.getChildren(id);
+    if (children.isNotEmpty) {
+      throw RepositoryException(
+        message: 'Cannot delete node with children',
+        code: RepositoryErrorCode.unknown,
+      );
+    }
+
     await _repository.delete(id);
   }
 
@@ -53,12 +64,37 @@ class NodeService {
     return await _repository.getChildren(id);
   }
 
-  Future<List<Node>> trace(String id) async {
-    return await _repository.trace(id);
+  Future<List<Node>> trace(String nodeId) async {
+    final List<Node> path = [];
+    String? currentId = nodeId;
+
+    while (currentId != null) {
+      try {
+        final node = await _repository.get(currentId);
+        path.add(node);
+        currentId = node.previous;
+      } catch (e) {
+        if (e is RepositoryException &&
+            e.code == RepositoryErrorCode.notFound) {
+          break;
+        }
+        rethrow;
+      }
+    }
+
+    return path;
   }
 
   Future<List<Node>> getSpatialNodes(String spatialPrefix) async {
     return await _repository.getSpatialNodes(spatialPrefix);
+  }
+
+  Future<List<Node>> getNodesByRoot(String rootId) async {
+    return await _repository.getByRoot(rootId);
+  }
+
+  Future<List<Node>> getAllNodes() async {
+    return await _repository.query();
   }
 
   void dispose() {
